fix: set default search limit for map page correctly
